requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
requests>=2.31.0
python-dotenv>=1.0.0
python-telegram-bot>=20.6
PyYAML>=6.0
